requirements.txt
pandas>=1.3.0
pulp>=2.4
openpyxl>=3.0.7
pyyaml>=5.4.1
